requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
decorator==5.1.1
executing==2.1.0
fastapi==0.115.5
google-api-core==2.23.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httptools==0.6.4
idna==3.10
marshmallow==3.23.1
numpy==2.1.3
oauthlib==3.2.2
parso==0.8.4
pexpect==4.9.0
piexif==1.1.3
pillow==10.4.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pygltflib==1.16.2
Pygments==2.18.0
PyMuPDF==1.24.14
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
stegano==0.11.4
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==14.1
wrapt==1.16.0
